add a button to switch the mip viewport
